+++ src/routes/Registration.tsx
@@ -1,6 +1,6 @@
 import { VStack } from '@chakra-ui/react';
 import { NavBar } from '../components/NavBar';
-import AttendeeInformation from '../components/Registration/Pages/AttendeeInformation';
+import AttendeeInformation from '../components/Registration/pages/AttendeeInformation';
 
 export default function Registration() {
 	return (
